relocates disabledKeypointFilterRules
