+++ web_ui/src/pages/media/utils.ts
@@ -413,3 +413,10 @@ export const filterPageMedias = (
         ),
     })),
 });
+
+export const disabledKeypointFilterRules = [
+    SearchRuleField.LabelId,
+    SearchRuleField.ShapeType,
+    SearchRuleField.ShapeAreaPixel,
+    SearchRuleField.ShapeAreaPercentage,
+];
